fix(host): convert migrate set downtime value to float64 (#15475)
* fix(host): convert migrate set downtime value to float64

* fix(host): live migration optimize

Qemu will send `STOP` event on start last time migration, but this
moment migration not completed. So we need wait magration complete.
Configurable Auto converge max cpu throttle.

+++ pkg/hostman/monitor/qmp.go
@@ -611,7 +611,7 @@ func (m *QmpMonitor) DeviceAdd(dev string, params map[string]string, callback St
 	// m.Query(cmd, cb)
 }
 
-func (m *QmpMonitor) MigrateSetDowntime(dtSec float32, callback StringCallback) {
+func (m *QmpMonitor) MigrateSetDowntime(dtSec float64, callback StringCallback) {
 	m.HumanMonitorCommand(fmt.Sprintf("migrate_set_downtime %f", dtSec), callback)
 }
